Add garage-ros-intera
The dockerfile for garage-ros-intera was added. It includes all the
required modifications to run the ROS Python2 libraries with Python3.
The images are based on the images built from the garage repository.
The corresponding docker-compose and makefile targets have been added,
as well as an update to the README for docker file.