MuJoCo: New Toy Environment (#23)
Implement a new environment for MuJoCo called ToyEnv. The env's
objective is to train Sawyer to pick up a peg, insert it into a toy
box lid with a hole, and open the box. Thanks to the sequential and
compositional nature of the environment, additional tasks can be
appended as desired.