Add garage-ros-intera
The dockerfile for garage-ros-intera was added. It includes all the 
required modifications to run the ROS Python2 libraries with Python3,
specifically, rospy, intera_interface, moveit_commander and 
moveit_msgs.
The images are based on the images built from the garage repository.
The corresponding docker-compose and makefile targets have been added,
as well as an update to the instructions in docker/README.md.
Finally, an example file to control Sawyer was added under the examples
folder.