Modify reward for pick task in Sawyer ToyEnv
Since the pick and place task is not working well, we're taking baby
steps to fix it. In this commit we have accomplished to grasp the peg
but not to lift it, and we made the following changes to do so:
- Reduce the action space to a cube of 10 cm per side.
- Include a gripper site (position placeholder) for the point when both
fingers in the gripper meet when closed.
- Consider two terms in the reward function: a binary reward of 100 when
the peg is grasped and 0 when it's not, and a penalization equivalent to
the distance between the gripper site and the peg.
- Never consider the task as done and keep feeding the reward until the
end (horizon) of the rollout.